requirements.txt
asn1crypto==1.3.0
bcrypt==3.1.7
blinker==1.4
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
coverage==5.0
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-vision==0.38.1
googleapis-common-protos==1.51.0
grpcio==1.23.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
olefile==0.46
Pillow==6.2.1
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.8
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==2.2.1
